Revert "refactor: extract chat stream recovery and sandbox lifecycle logic fr…"
This reverts commit 59d0654f85a55b06a8a87c3c24db8cba13bc1f2f.
